+++ filter_struct_of_vectors.cpp
@@ -14,7 +14,7 @@ struct S
 int main()
 {
   S s;
-  auto size = end_data() - begin_data();
+  auto size = static_cast<size_t>(end_data() - begin_data());
   s.xs.reserve(size);
   s.ys.reserve(size);
   s.zs.reserve(size);
